requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
langchain-text-splitters>=0.0.1
chromadb>=0.4.18
pypdf>=3.17.4
sentence-transformers>=2.2.2
transformers>=4.34.0
torch>=2.1.0
accelerate>=0.20.0
sentencepiece>=0.1.99
safetensors>=0.3.0
python-dotenv>=1.0.0